Pull JsonMutantLocation and JsonMutantPosition to their own files
